src/main/java/hello/HelloController.java
package hello;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.bind.annotation.RequestMapping;

@RestController
public class HelloController {

    @Autowired
    SayHelloService service;

    public HelloController(SayHelloService service){
        this.service = service;
    }

    @RequestMapping("/")
    public String index() {
        return "Greetings from Spring Boot!";
    }

    @RequestMapping("/hello")
    public String sayHello(@RequestParam("name") String name){
        return  service.sayHello(name.toUpperCase());

    }

    
}
